+++ packages/7715-permission-types/src/permissions/caveats/erc20TokenPeriodic.ts
@@ -33,7 +33,7 @@ export function makeErc20TokenPeriodicDecoderConfig(
 ): PermissionDecoderConfig {
   const {
     timestampEnforcer,
-    erc20PeriodicEnforcer,
+    erc20PeriodTransferEnforcer,
     valueLteEnforcer,
     nonceEnforcer,
     allowedCalldataEnforcer,
@@ -49,7 +49,7 @@ export function makeErc20TokenPeriodicDecoderConfig(
       allowedCalldataEnforcer, // payee rule
     ],
     requiredEnforcers: {
-      [erc20PeriodicEnforcer]: 1,
+      [erc20PeriodTransferEnforcer]: 1,
       [valueLteEnforcer]: 1,
       [nonceEnforcer]: 1,
     },
@@ -69,7 +69,7 @@ function validateAndDecodeData(
   caveats: ChecksumCaveat[],
   contractAddresses: ChecksumEnforcersByChainId,
 ): DecodedPermissionData<Erc20TokenPeriodicPermission> {
-  const { erc20PeriodicEnforcer, valueLteEnforcer } = contractAddresses;
+  const { erc20PeriodTransferEnforcer, valueLteEnforcer } = contractAddresses;
 
   const valueLteTerms = getTermsByEnforcer({
     caveats,
@@ -81,7 +81,7 @@ function validateAndDecodeData(
 
   const terms = getTermsByEnforcer({
     caveats,
-    enforcer: erc20PeriodicEnforcer,
+    enforcer: erc20PeriodTransferEnforcer,
   });
 
   const {
@@ -128,7 +128,7 @@ function validateAndDecodeData(
  */
 export type Erc20TokenPeriodicEnforcers = Pick<
   ChecksumEnforcersByChainId,
-  'erc20PeriodicEnforcer' | 'valueLteEnforcer'
+  'erc20PeriodTransferEnforcer' | 'valueLteEnforcer'
 >;
 
 /**
@@ -175,7 +175,7 @@ export function createErc20TokenPeriodicCaveats({
   }
 
   const erc20PeriodCaveat: Caveat = {
-    enforcer: contracts.erc20PeriodicEnforcer,
+    enforcer: contracts.erc20PeriodTransferEnforcer,
     terms: createERC20TokenPeriodTransferTerms({
       tokenAddress,
       periodAmount: periodAmountBigInt,