+++ packages/7715-permission-types/test/permissions/caveats/erc20TokenAllowance.test.ts
@@ -31,7 +31,7 @@ describe('erc20-token-allowance decoder config', () => {
   const contracts = DELEGATOR_CONTRACTS['1.3.0'][chainId];
   const {
     timestampEnforcer,
-    erc20PeriodicEnforcer,
+    erc20PeriodTransferEnforcer,
     valueLteEnforcer,
     nonceEnforcer,
     allowedCalldataEnforcer,
@@ -52,7 +52,7 @@ describe('erc20-token-allowance decoder config', () => {
     valueLteTerms: Hex = ZERO_32_BYTES,
   ): ChecksumCaveat[] => [
     {
-      enforcer: erc20PeriodicEnforcer,
+      enforcer: erc20PeriodTransferEnforcer,
       terms: erc20PeriodicTerms,
       args: '0x',
     },
@@ -71,7 +71,7 @@ describe('erc20-token-allowance decoder config', () => {
   describe('static configuration', () => {
     it('exposes expected required enforcers', () => {
       expect(decoder.requiredEnforcers).toStrictEqual({
-        [erc20PeriodicEnforcer]: 1,
+        [erc20PeriodTransferEnforcer]: 1,
         [valueLteEnforcer]: 1,
         [nonceEnforcer]: 1,
       });
@@ -177,7 +177,7 @@ describe('createErc20TokenAllowanceCaveats()', () => {
   const startTime = 1729900800;
 
   const contracts: Erc20TokenAllowanceEnforcers = {
-    erc20PeriodicEnforcer: '0x7356Ed4321Ff9e7DAE246461829cDC170ff660Ab',
+    erc20PeriodTransferEnforcer: '0x7356Ed4321Ff9e7DAE246461829cDC170ff660Ab',
     valueLteEnforcer: '0x5e12Ca712176E7557e4fAa1c8cc27382B60B5e39',
   };
 
@@ -201,7 +201,7 @@ describe('createErc20TokenAllowanceCaveats()', () => {
 
     expect(caveats).toStrictEqual([
       {
-        enforcer: contracts.erc20PeriodicEnforcer,
+        enforcer: contracts.erc20PeriodTransferEnforcer,
         terms: expectedTerms,
         args: '0x',
       },
@@ -302,7 +302,7 @@ describe('createErc20TokenAllowanceCaveats()', () => {
     });
     const erc20AllowanceTerms = caveats[0]?.terms as Hex;
 
-    expect(caveats[0]?.enforcer).toBe(contracts.erc20PeriodicEnforcer);
+    expect(caveats[0]?.enforcer).toBe(contracts.erc20PeriodTransferEnforcer);
     expect(
       erc20AllowanceTerms.startsWith(`0x${alternateTokenAddress.slice(2)}`),
     ).toBe(true);