+++ packages/7715-permission-types/test/permissions/caveats/nativeTokenAllowance.test.ts
@@ -30,7 +30,7 @@ describe('native-token-allowance decoder config', () => {
   const contracts = DELEGATOR_CONTRACTS['1.3.0'][chainId];
   const {
     timestampEnforcer,
-    nativeTokenPeriodicEnforcer,
+    nativeTokenPeriodTransferEnforcer,
     exactCalldataEnforcer,
     nonceEnforcer,
     allowedTargetsEnforcer,
@@ -51,7 +51,7 @@ describe('native-token-allowance decoder config', () => {
     exactCalldataTerms: Hex = '0x',
   ): ChecksumCaveat[] => [
     {
-      enforcer: nativeTokenPeriodicEnforcer,
+      enforcer: nativeTokenPeriodTransferEnforcer,
       terms: nativeTokenPeriodicTerms,
       args: '0x',
     },
@@ -70,7 +70,7 @@ describe('native-token-allowance decoder config', () => {
   describe('static configuration', () => {
     it('exposes expected required enforcers', () => {
       expect(decoder.requiredEnforcers).toStrictEqual({
-        [nativeTokenPeriodicEnforcer]: 1,
+        [nativeTokenPeriodTransferEnforcer]: 1,
         [exactCalldataEnforcer]: 1,
         [nonceEnforcer]: 1,
       });
@@ -174,7 +174,8 @@ describe('createNativeTokenAllowanceCaveats()', () => {
   const startTime = 1729900800;
 
   const contracts: NativeTokenAllowanceEnforcers = {
-    nativeTokenPeriodicEnforcer: '0x7356Ed4321Ff9e7DAE246461829cDC170ff660Ab',
+    nativeTokenPeriodTransferEnforcer:
+      '0x7356Ed4321Ff9e7DAE246461829cDC170ff660Ab',
     exactCalldataEnforcer: '0x5e12Ca712176E7557e4fAa1c8cc27382B60B5e39',
   };
 
@@ -197,7 +198,7 @@ describe('createNativeTokenAllowanceCaveats()', () => {
 
     expect(caveats).toStrictEqual([
       {
-        enforcer: contracts.nativeTokenPeriodicEnforcer,
+        enforcer: contracts.nativeTokenPeriodTransferEnforcer,
         terms: expectedTerms,
         args: '0x',
       },