+++ packages/7715-permission-types/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Incorrectly named caveats used in caveat factory functions ([#274](https://github.com/MetaMask/smart-accounts-kit/pull/274))
+  - erc20PeriodicEnforcer is now erc20PeriodTransferEnforcer
+  - nativeTokenPeriodicEnforcer is now nativeTokenPeriodTransferEnforcer
+
 ## [0.8.0]
 
 ### Added